requirements/heroku.txt
-r base.txt

dj-database-url==0.2.1
django-extensions==1.1.1
django-storages==1.1.8
